@@ -4,27 +4,20 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.validation.constraints.Min;
-
-import org.hibernate.annotations.BatchSize;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -45,40 +38,38 @@ public class Book {
 	private Long bookId;
 	private String bookImage;
 	private String title;
-	@ManyToMany(fetch = FetchType.LAZY , cascade = CascadeType.ALL)
+	@ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
 	@JoinTable(name = "book_languages", joinColumns = @JoinColumn(name = "book_id"), inverseJoinColumns = @JoinColumn(name = "language_id"))
 	private List<Language> languages = new ArrayList<>();
 	private LocalDate publicationDate;
 	private String authorName;
-	
+
 	private int edition;
-	
-	    @Column(length = 30)  // Use a unique name for the column
+
+	@Column(length = 30) // Use a unique name for the column
 	private boolean isAvailable = true;
 	private int stock;
 	@Enumerated(EnumType.STRING)
-    @Column(length = 30, name = "genre")  // Use a unique name for the column
-    private Genre genre;
+	@Column(length = 30, name = "genre") // Use a unique name for the column
+	private Genre genre;
 	@OneToMany(mappedBy = "book", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
 	private List<Review> reviews = new ArrayList<Review>();
-	@ManyToOne
-	@JoinColumn(name = "cart_id")
-	private Cart cart;
-	@ManyToOne
-	@JoinColumn(name = "order_id")
-	private Order order; // Represents the order this book belongs to
-	
+	@ManyToMany(mappedBy = "books", fetch = FetchType.LAZY)
+	private List<Cart> carts = new ArrayList<>();
+//	@ManyToOne
+//	@JoinColumn(name = "order_id")
+//	private Order order; // Represents the order this book belongs to
+
 	private float markedPrice;
-	  private float sellingPrice;
-	
+	private float sellingPrice;
+
 	public Book() {
 		super();
 	}
 
-
 	public Book(Long bookId, String bookImage, String title, List<Language> languages, LocalDate publicationDate,
 			String authorName, int edition, boolean isAvailable, int stock, Genre genre, List<Review> reviews,
-			Cart cart, Order order, float markedPrice, float sellingPrice) {
+			List<Cart> carts, float markedPrice, float sellingPrice) {
 		super();
 		this.bookId = bookId;
 		this.bookImage = bookImage;
@@ -91,10 +82,21 @@ public class Book {
 		this.stock = stock;
 		this.genre = genre;
 		this.reviews = reviews;
-		this.cart = cart;
-		this.order = order;
+		this.carts = carts;
+//		this.order = order;
 		this.markedPrice = markedPrice;
 		this.sellingPrice = sellingPrice;
+	}
+
+
+	public Book(Long bookId, String title, String bookImage, float sellingPrice, Genre genre, int edition, String authorName) {
+		// TODO Auto-generated constructor stub
+		this.bookId = bookId;
+		this.bookImage = bookImage;
+		this.title = title;
+		this.edition = edition;
+		this.genre = genre;
+		this.authorName = authorName;
 	}
 
 	public Long getBookId() {
@@ -184,21 +186,31 @@ public class Book {
 	public void setReviews(List<Review> reviews) {
 		this.reviews = reviews;
 	}
+	
+	
 
-	public Cart getCart() {
-		return cart;
+//	public Cart getCart() {
+//		return cart;
+//	}
+//
+//	public void setCart(Cart cart) {
+//		this.cart = cart;
+//	}
+
+//	public Order getOrder() {
+//		return order;
+//	}
+//
+//	public void setOrder(Order order) {
+//		this.order = order;
+//	}
+
+	public List<Cart> getCarts() {
+		return carts;
 	}
 
-	public void setCart(Cart cart) {
-		this.cart = cart;
-	}
-
-	public Order getOrder() {
-		return order;
-	}
-
-	public void setOrder(Order order) {
-		this.order = order;
+	public void setCarts(List<Cart> carts) {
+		this.carts = carts;
 	}
 
 	public float getMarkedPrice() {
@@ -216,8 +228,5 @@ public class Book {
 	public void setSellingPrice(float sellingPrice) {
 		this.sellingPrice = sellingPrice;
 	}
-	
-	
-	
-	
+
 }
