@@ -4,24 +4,21 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-import javax.validation.constraints.Pattern;
-
-import lombok.Getter;
-import lombok.Setter;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
+import jakarta.validation.constraints.Pattern;
 import lombok.ToString;
 
 /**
@@ -33,13 +30,6 @@ import lombok.ToString;
 @ToString
 @Table(name = "users")
 public class User {
-	public Cart getCart() {
-		return cart;
-	}
-
-	public void setCart(Cart cart) {
-		this.cart = cart;
-	}
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -55,7 +45,7 @@ public class User {
 	private String confirmPassword;
 	private LocalDate dateOfBirth;
 	private String addressLine1;
-	private String addressLin0e2;
+	private String addressLine2;
 	// Many-to-one relationship with City using zip code
 	@ManyToOne()
 	@JoinColumn(name = "city", referencedColumnName = "cityId")
@@ -72,8 +62,9 @@ public class User {
 	@OneToMany(mappedBy = "user")
 	private List<Order> orders = new ArrayList<>();
 	
-	 @OneToOne( cascade = CascadeType.ALL, orphanRemoval = true)
-	    private Cart cart;
+	@OneToOne(cascade = CascadeType.ALL, orphanRemoval = true)
+    @JoinColumn(name = "cart_id") // This column should be present in the users table
+    private Cart cart;
 
 	public User() {
 		super();
@@ -93,12 +84,19 @@ public class User {
 		this.confirmPassword = confirmPassword;
 		this.dateOfBirth = dateOfBirth;
 		this.addressLine1 = addressLine1;
-		this.addressLin0e2 = addressLin0e2;
+		this.addressLine2 = addressLin0e2;
 		this.city = city;
 		this.phoneNumber = phoneNumber;
 		this.role = role;
 		this.reviews = reviews;
 		this.orders = orders;
+	}
+	public Cart getCart() {
+		return cart;
+	}
+	
+	public void setCart(Cart cart) {
+		this.cart = cart;
 	}
 
 	public Long getUserId() {
@@ -173,12 +171,12 @@ public class User {
 		this.addressLine1 = addressLine1;
 	}
 
-	public String getAddressLin0e2() {
-		return addressLin0e2;
+	public String getAddressLine2() {
+		return addressLine2;
 	}
 
 	public void setAddressLin0e2(String addressLin0e2) {
-		this.addressLin0e2 = addressLin0e2;
+		this.addressLine2 = addressLin0e2;
 	}
 
 	public City getCity() {
